WhisperX_chinese_2_speakers/requirements.txt
whisperx>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
soundfile>=0.12.1 